Code modernisation and concurrency fixes. Still have racing conditions in the archiver module :(

+++ DataController.h
@@ -63,16 +63,16 @@ struct MVNodeSaver;
 {
   MVColoumns *          coloumns;
   NSDictionary *        attributes;
-  uint32_t              offset;           // for sorting if necessary
-  uint32_t              coloumnsOffset;   // offset of coloumns
-  uint32_t              attributesOffset; // offset of attribues
+  uint64_t              offset;           // for sorting if necessary
+  uint64_t              coloumnsOffset;   // offset of coloumns
+  uint64_t              attributesOffset; // offset of attribues
   BOOL                  deleted;
   BOOL                  dirty;            // eg. attributes has changed
 }
 
 @property (nonatomic)   NSDictionary * attributes;
 @property (nonatomic)   MVColoumns * coloumns;
-@property (nonatomic)   uint32_t offset;
+@property (nonatomic)   uint64_t offset;
 @property (nonatomic)   BOOL deleted;
 @property (nonatomic)   BOOL dirty;
 
@@ -99,7 +99,7 @@ struct MVNodeSaver;
 
 - (void)                popRow;
 - (void)                appendRow:(id)col0 :(id)col1 :(id)col2 :(id)col3;
-- (void)                insertRowWithOffset:(uint32_t)offset :(id)col0 :(id)col1 :(id)col2 :(id)col3;
+- (void)                insertRowWithOffset:(uint64_t)offset :(id)col0 :(id)col1 :(id)col2 :(id)col3;
 - (void)                updateCellContentTo:(id)object atRow:(NSUInteger)rowIndex andCol:(NSUInteger)colIndex;
 
 - (NSUInteger)          rowCount;
@@ -118,7 +118,7 @@ struct MVNodeSaver;
   NSRange               dataRange;
   MVTable *             details;
   NSMutableDictionary * userInfo;
-  uint32_t              detailsOffset;
+  uint64_t              detailsOffset;
 }
 
 @property (nonatomic)           NSString *            caption;
@@ -126,12 +126,12 @@ struct MVNodeSaver;
 @property (nonatomic)           NSRange               dataRange;
 @property (nonatomic)           MVTable *             details;
 @property (nonatomic)           NSMutableDictionary * userInfo;
-@property (nonatomic)           uint32_t              detailsOffset;
+@property (nonatomic)           uint64_t              detailsOffset;
 
 - (NSUInteger)          numberOfChildren;
 - (MVNode *)            childAtIndex:(NSUInteger)n;
-- (MVNode *)            insertChild:(NSString *)_caption location:(uint32_t)location length:(uint32_t)length;
-- (MVNode *)            insertChildWithDetails:(NSString *)_caption location:(uint32_t)location length:(uint32_t)length saver:(MVNodeSaver &)saver;
+- (MVNode *)            insertChild:(NSString *)_caption location:(uint64_t)location length:(uint64_t)length;
+- (MVNode *)            insertChildWithDetails:(NSString *)_caption location:(uint64_t)location length:(uint64_t)length saver:(MVNodeSaver &)saver;
 - (MVNode *)            findNodeByUserInfo:(NSDictionary *)uinfo;
 - (void)                openDetails;  // open swap file for reading details on demand
 - (void)                closeDetails; // close swap file
@@ -165,7 +165,7 @@ struct MVNodeSaver;
 -(NSString *)           getMachine:(cpu_type_t)cputype;
 -(NSString *)           getARMCpu:(cpu_subtype_t)cpusubtype;
 
-- (void)                createLayouts:(MVNode *)parent location:(uint32_t)location length:(uint32_t)length;
+- (void)                createLayouts:(MVNode *)parent location:(uint64_t)location length:(uint64_t)length;
 - (void)                updateTreeView: (MVNode *)node;
 - (void)                updateTableView;
 - (void)                updateStatus: (NSString *)status;
